Make attribute drag and drop compatible with SWADE 3.2
The attribute drag and drop feature was modified to reduce complexity and be compatible with 3.2. The code related to handling attribute labels' dragstart event was removed from the attribute_card.js file and replaced in the brsw2-init.js file with the creation of an Attribute macro. This macro simplifies the logic of whether to display a Better Rolls 2 card or a default card, depending on the event and settings.